+++ CHANGELOG.md
@@ -2,8 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2020-05-14
+
+### Added
+
+- Added [`User::hasViewed`](docs/User.md#userhasviewed)
+- Added [`User::setViewed`](docs/User.md#usersetviewed)
+
+### Breaking changes
+
+- Updated [`User::addExperiment`](docs/User.md#useraddexperiment) to include a `$viewed` argument before `$converted`
+- Renamed `Group::getSize` to [`Group::getViews`](docs/Group.md#groupgetviews)
+- Renamed `Group::setSize` to [`Group::setViews`](docs/Group.md#groupsetviews)
+- Renamed `Experiment::getCoverage` to [`Experiment::getAllocation`](docs/Experiment.md#experimentgetallocation)
+- Renamed `Experiment::setCoverage` to [`Experiment::setAllocation`](docs/Experiment.md#experimentsetallocation)
+
 ## [1.0.0] - 2020-04-29
 
 Initial release
 
+[1.1.0]: https://github.com/andreekeberg/abby/releases/tag/1.1.0
 [1.0.0]: https://github.com/andreekeberg/abby/releases/tag/1.0.0